# swagger.yaml
openapi: '3.0.0'
info:
  title: E-Commerce-API
  version: '1.0'
servers:
  - url: https://itwashard-auufff.onrender.com/api
paths:
...
  /products:
    post:
      tags:
        - Products
      summary: Create Product
      operationId: CreateProduct
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/CreateProductRequest'
            example:
              name: accent chair
              price: 25999
              image: /uploads/computer-6.jpeg
              colors:
                - '#ff0000'
                - '#00ff00'
                - '#0000ff'
              company: marcos
              description: Cloud bread VHS hell of banjo bicycle rights jianbing umami mumblecore etsy 8-bit pok pok +1 wolf. Vexillologist yr dreamcatcher waistcoat, authentic chillwave trust fund. Viral typewriter fingerstache pinterest pork belly narwhal. Schlitz venmo everyday carry kitsch pitchfork chillwave iPhone taiyaki trust fund hashtag kinfolk microdosing gochujang live-edge
              category: office
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
    get:
      tags:
        - Products
      summary: Get All Products
      operationId: GetAllProducts
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
...
  /reviews:
    get:
      tags:
        - Reviews
      summary: Get All Reviews
      operationId: GetAllReviews
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
    post:
      tags:
        - Reviews
      summary: Create Review
      operationId: CreateReview
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/CreateReviewRequest'
            example:
              product: 65a3ab6c1ee99207f26bfddc
              rating: 2
              title: 'Thats bad '
              comment: No need to buy this thing
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
...
components:
  schemas:
...
    CreateProductRequest:
      title: CreateProductRequest
      required:
        - name
        - price
        - image
        - colors
        - company
        - description
        - category
      type: object
      properties:
        name:
          type: string
        price:
          type: integer
        image:
          type: string
        colors:
          type: array
          items:
            type: string
          description: ''
        company:
          type: string
        description:
          type: string
        category:
          type: string
      example:
        name: accent chair
        price: 25999
        image: /uploads/computer-6.jpeg
        colors:
          - '#ff0000'
          - '#00ff00'
          - '#0000ff'
        company: marcos
        description: Cloud bread VHS hell of banjo bicycle rights jianbing umami mumblecore etsy 8-bit pok pok +1 wolf. Vexillologist yr dreamcatcher waistcoat, authentic chillwave trust fund. Viral typewriter fingerstache pinterest pork belly narwhal. Schlitz venmo everyday carry kitsch pitchfork chillwave iPhone taiyaki trust fund hashtag kinfolk microdosing gochujang live-edge
        category: office
    CreateReviewRequest:
      title: CreateReviewRequest
      required:
        - product
        - rating
        - title
        - comment
      type: object
      properties:
        product:
          type: string
        rating:
          type: integer
        title:
          type: string
        comment:
          type: string
      example:
        product: 65a3ab6c1ee99207f26bfddc
        rating: 2
        title: 'Thats bad '
        comment: No need to buy this thing
...